Move txstore package to legacy directory
