+++ legacy/txstore/tx_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcd/wire"
 	"github.com/btcsuite/btcutil"
-	. "github.com/btcsuite/btcwallet/txstore"
+	. "github.com/btcsuite/btcwallet/legacy/txstore"
 )
 
 // Received transaction output for mainnet outpoint
